Add map to report
