+++ monkey_island/cc/ui/src/components/map/preview-pane/PreviewPane.js
